refactor: Utils class 추가 - getRandomNumber 메서드 : 서로 다른 랜덤 숫자 3개 생성 - convertStringToInt 메서드 : string을 int array로 변환
